chore: typed return for client factory
