+++ tests/conformance/test_conformance_tools.py
@@ -7,7 +7,7 @@ from openai.types.beta.threads.text import Text
 from openai.types.beta.threads.message import Message
 import re
 
-from .utils import client_config_factory, text_file_path
+from ..utils.client import client_config_factory, text_file_path
 
 
 def make_vector_store_with_file(client):
@@ -46,10 +46,10 @@ def validate_annotation_format(annotation):
 @pytest.mark.parametrize("client_name", ["openai", "leapfrogai"])
 def test_thread_file_annotations(client_name):
     config = client_config_factory(client_name)
-    client = config["client"]  # shorthand
+    client = config.client  # shorthand
 
     vector_store = make_vector_store_with_file(client)
-    assistant = make_test_assistant(client, config["model"], vector_store.id)
+    assistant = make_test_assistant(client, config.model, vector_store.id)
     thread = client.beta.threads.create()
 
     client.beta.threads.messages.create(
